pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "slate"
version = "0.6.0"
description = "Extract text from PDF documents easily"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT License"}
authors = [{name = "Tim McNamara", email = "paperless@timmcnamara.co.nz"}]
keywords = ["pdf", "text", "text-extraction"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing",
    "Topic :: Utilities",
]
urls = {repository = "http://github.com/timClicks/slate"}
dependencies = ["pdfminer.six"]

[tool.pytest.ini_options]
addopts = "--doctest-modules"

[tool.mypy]
exclude = "tests"
show_error_codes = true
ignore_missing_imports = true
warn_unreachable = true
strict = true
